client_type

+++ cloudimgWinform/dao/UploadTaskDao.cs
@@ -39,7 +39,7 @@ namespace cloudimgWinform.dao
                 //初始化数据表和库
                 if (first)
                 {
-                    string sql = "create table t_uploadtask (id integer primary key autoincrement,name varchar(100), path varchar(256),upload_path varchar(256),convert_path varchar(256),associated_img_path varhar(256), size long,md5 varchar(32),scan_rate int,width long,height long,resolution float,status int)";
+                    string sql = "create table t_uploadtask (id integer primary key autoincrement,name varchar(100), path varchar(256),upload_path varchar(256),convert_path varchar(256),associated_img_path varhar(256), size long,md5 varchar(32),scan_rate int,width long,height long,resolution float,status int,user_id integer)";
                     SQLiteCommand command=SQLiteHelper.CreateCommand(connection, sql, null);
                     command.ExecuteNonQuery();
                 }
@@ -67,13 +67,14 @@ namespace cloudimgWinform.dao
 
         public static bool addTask(UploadTask ut)
         {
-            String sql = "insert into t_uploadtask(name,path,size,md5,status) values(@name,@path,@size,@md5,@status)";
+            String sql = "insert into t_uploadtask(name,path,size,md5,status,user_id) values(@name,@path,@size,@md5,@status,@user_id)";
             SQLiteParameter p1= SQLiteHelper.CreateParameter("name", DbType.String, ut.name);
             SQLiteParameter p2 = SQLiteHelper.CreateParameter("path", DbType.String, ut.path);
             SQLiteParameter p3 = SQLiteHelper.CreateParameter("size", DbType.Int64, ut.size);
             SQLiteParameter p4 = SQLiteHelper.CreateParameter("md5", DbType.String, ut.md5);
             SQLiteParameter p5= SQLiteHelper.CreateParameter("status", DbType.Int16, ut.status);
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1, p2, p3, p4,p5 });
+            SQLiteParameter p6 = SQLiteHelper.CreateParameter("user_id", DbType.Int16,User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1, p2, p3, p4,p5, p6 });
             int result=command.ExecuteNonQuery();
             Console.WriteLine("保存数据" + result + "条");
             return result>0?true:false;
@@ -91,9 +92,10 @@ namespace cloudimgWinform.dao
 
         public static bool delTaskByStatus(int status)
         {
-            String sql = "delete from t_uploadtask where status=@status";
+            String sql = "delete from t_uploadtask where status=@status and user_id=@user_id";
             SQLiteParameter p1 = SQLiteHelper.CreateParameter("status", DbType.Int32, status);
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1 });
+            SQLiteParameter p2 = SQLiteHelper.CreateParameter("user_id", DbType.Int16, User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1,p2 });
             int result = command.ExecuteNonQuery();
             Console.WriteLine("删除数据" + result + "条");
             return result > 0 ? true : false;
@@ -101,8 +103,9 @@ namespace cloudimgWinform.dao
 
         public static bool delAllTask()
         {
-            String sql = "delete from t_uploadtask";
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql,null);
+            String sql = "delete from t_uploadtask where user_id=@user_id";
+            SQLiteParameter p1 = SQLiteHelper.CreateParameter("user_id", DbType.Int16, User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql,new SQLiteParameter[] { p1});
             int result = command.ExecuteNonQuery();
             Console.WriteLine("删除数据" + result + "条");
             return result > 0 ? true : false;
@@ -110,8 +113,9 @@ namespace cloudimgWinform.dao
 
         public static IList<UploadTask> query()
         {
-            String sql = "select * from t_uploadtask";
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql,null);
+            String sql = "select * from t_uploadtask where user_id=@user_id";
+            SQLiteParameter p1 = SQLiteHelper.CreateParameter("user_id", DbType.Int16, User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1 });
             SQLiteDataReader reader = command.ExecuteReader();
             UploadTask task=null;
             IList<UploadTask> tasks = new BindingList<UploadTask>();
@@ -129,9 +133,10 @@ namespace cloudimgWinform.dao
         public static List<UploadTask> getByStatus(int status)
         {
             List<UploadTask> tasks = new List<UploadTask>();
-            String sql = "select * from t_uploadtask where status=@status order by id limit 10";
+            String sql = "select * from t_uploadtask where status=@status and user_id=@user_id order by id limit 10";
             SQLiteParameter p1 = SQLiteHelper.CreateParameter("status", DbType.Int32, status);
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1 });
+            SQLiteParameter p2 = SQLiteHelper.CreateParameter("user_id", DbType.Int16, User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1,p2 });
             SQLiteDataReader reader = command.ExecuteReader();
             UploadTask task = null;
             while (reader.Read())
@@ -145,9 +150,10 @@ namespace cloudimgWinform.dao
 
         public static UploadTask getByMd5(String md5)
         {
-            String sql = "select * from t_uploadtask where md5=@md5 limit 1";
+            String sql = "select * from t_uploadtask where md5=@md5 and user_id=user_id limit 1";
             SQLiteParameter p1 = SQLiteHelper.CreateParameter("md5", DbType.String, md5);
-            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1 });
+            SQLiteParameter p2 = SQLiteHelper.CreateParameter("user_id", DbType.Int16, User.loginUser.userId);
+            SQLiteCommand command = SQLiteHelper.CreateCommand(connection, sql, new SQLiteParameter[] { p1,p2 });
             SQLiteDataReader reader = command.ExecuteReader();
             UploadTask task = null;
             while (reader.Read())
@@ -163,6 +169,7 @@ namespace cloudimgWinform.dao
             task.id = int.Parse(reader["id"].ToString());
             task.name = reader["name"].ToString();
             task.path = reader["path"].ToString();
+            task.md5 = reader["md5"].ToString();
             task.uploadPath = reader["upload_path"].ToString();
             task.associatedImgPath = reader["associated_img_path"].ToString();
             task.convertPath = reader["convert_path"].ToString();
